Work towards basic dev environment without docker
