+++ docker-net/.env
@@ -1,47 +1,6 @@
-# Copyright (c) Jupyter Development Team.
-# Distributed under the terms of the Modified BSD License.
+SSL_CERT=./secrets/jupyterhub.crt
+SSL_KEY=./secrets/jupyterhub.key
 
-# Use this file to set default values for environment variables specified in
-# docker-compose configuration file.  docker-compose will substitute these
-# values for environment variables in the configuration file IF the variables
-# are not set in the shell environment.
+SQLITE_DB=./stochssdb.sqlite
+DB_CONNECT=sqlite:///./stochssdb.sqlite
 
-# To override these values, set the shell environment variables.
-JUPYTERHUB_VERSION=0.9.6
-
-# Name of Docker machine
-DOCKER_MACHINE_NAME=jupyterhub
-
-# Name of Docker network
-DOCKER_NETWORK_NAME=jupyterhub-network
-
-# Single-user Jupyter Notebook server container image
-DOCKER_NOTEBOOK_IMAGE=jupyter/minimal-notebook
-
-# the local image we use, after pinning jupyterhub version
-LOCAL_NOTEBOOK_IMAGE=jupyterhub-user
-
-# Notebook directory in the container.
-# This will be /home/jovyan/work if the default
-# This directory is stored as a docker volume for each user
-DOCKER_NOTEBOOK_DIR=/home/jovyan/work
-
-# Docker run command to use when spawning single-user containers
-DOCKER_SPAWN_CMD=start-singleuser.sh
-
-# Name of JupyterHub container data volume
-DATA_VOLUME_HOST=jupyterhub-data
-
-# Data volume container mount point
-DATA_VOLUME_CONTAINER=/data
-
-# Name of JupyterHub postgres database data volume
-DB_VOLUME_HOST=jupyterhub-db-data
-
-# Postgres volume container mount point
-DB_VOLUME_CONTAINER=/var/lib/postgresql/data
-
-# The name of the postgres database containing JupyterHub state
-JUPYTERHUB_DB=jupyterhub
-
-STOCHSS_DB=models
