+++ docker-net/Pipfile
@@ -6,6 +6,12 @@ verify_ssl = true
 [dev-packages]
 
 [packages]
+jupyterhub = "==1.0"
+psycopg2 = "==2.7.7"
+sqlalchemy = "==1.2.18"
+python-dotenv = "*"
+gillespy2 = "*"
+notebook = "*"
 
 [requires]
 python_version = "3.6"
