--- docker-net/scripts/Pipfile
@@ -1,15 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-
-[packages]
-jupyterhub = "==0.9.4"
-psycopg2 = "==2.7.7"
-sqlalchemy = "==1.2.18"
-python-dotenv = "*"
-
-[requires]
-python_version = "3.6"
