--- docker-net/stochss/index.html
@@ -1,9 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="UTF-8">
-    <title>StochSS [DEV]</title>
-  </head>
-  <body>
-  <script type="text/javascript" src="app.bundle.js"></script></body>
-</html>
