+++ docker-net/stochss/static/index.html
@@ -1,0 +1,9 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="UTF-8">
+    <title>StochSS [DEV]</title>
+  </head>
+  <body>
+  <script type="text/javascript" src="app.bundle.js"></script></body>
+</html>
