+++ src/models/models.js
@@ -1,9 +1,9 @@
+var app = require('ampersand-app');
 var RestCollection = require('ampersand-rest-collection');
 var Model = require('./model');
 var path = require('path');
-var config = require('../config.js')(process.env.NODE_ENV);
 
 module.exports = RestCollection.extend({
   model: Model,
-  url: function () { return path.resolve(config.apiUrl, 'models') }
+  url: function () { return path.resolve(app.config.routePrefix, app.config.apiUrl, 'models') }
 });
