+++ webpack.config.js
@@ -7,7 +7,7 @@ module.exports = {
   entry: './src/app.js',
   output: {
     filename: 'app.bundle.js',
-    path: path.resolve(__dirname, "dist")
+    path: path.resolve(__dirname, "docker-net/stochss/static")
   },
   devtool: 'inline-source-map',
   devServer: {
